requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
bidict==0.22.1
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.8.0
dnspython==2.4.2
eventlet==0.34.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
frozenlist==1.4.0
gevent==23.9.1
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-texttospeech==2.14.2
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.2
grpcio==1.59.3
grpcio-status==1.59.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.28.0
packaging==23.2
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pygame==2.5.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
rsa==4.9
serverless-wsgi==3.0.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
watchdog==3.0.0
websocket-client==1.7.0
Werkzeug==3.0.1
wsproto==1.2.0
yarl==1.9.3
zope.event==5.0
zope.interface==6.1
